refactor(//core/conversion/arg): Pull out arg class from converters so they can be used for evaluators too.
