+++ core/conversion/arg/BUILD
@@ -8,32 +8,32 @@ config_setting(
 )
 
 cc_library(
-    name = "conversion",
+    name = "arg",
     hdrs = [
-        "conversion.h",
+        "Arg.h",
+        "Arg_inl.h"
     ],
     srcs = [
-        "conversion.cpp",
-        "conversion_blacklist.cpp",
-        "InterfaceTypes.cpp"
+        "Arg.cpp",
     ],
     deps = [
         "@tensorrt//:nvinfer",
-        "//core/conversion/arg",
+        "//core/util:prelude",
         "//core/conversion/conversionctx",
-        "//core/conversion/converters",
-        "//core/conversion/evaluators",
-        "//core/util:prelude"
     ] + select({
         ":use_pre_cxx11_abi":  ["@libtorch_pre_cxx11_abi//:libtorch"],
         "//conditions:default":  ["@libtorch//:libtorch"],
     }),
+    alwayslink = True,
 )
 
 load("@rules_pkg//:pkg.bzl", "pkg_tar")
 
 pkg_tar(
     name = "include",
-    package_dir = "core/conversion/",
-    srcs = ["conversion.h"],
+    package_dir = "core/conversion/arg/",
+    srcs = [
+        "Arg.h",
+        "Arg_inl.h"
+    ],
 )
